+++ nym-vpn-android/app/src/main/java/net/nymtech/nymvpn/data/datastore/DataStoreSettingsRepository.kt
@@ -33,7 +33,8 @@ class DataStoreSettingsRepository(private val dataStoreManager: DataStoreManager
 	private val sentryEnabled = booleanPreferencesKey("SENTRY_ENABLED")
 	private val statsEnabled = booleanPreferencesKey("STATISTICS_ENABLED")
 	private val statsDialogSkip = booleanPreferencesKey("STATISTICS_DIALOG_SKIP")
-	private val welcomeScreenCompleted = booleanPreferencesKey("WELCOME_SCREEN_COMPLETE")
+	private val onboardingCompleted = booleanPreferencesKey("ONBOARDING_COMPLETE")
+	private val technicalOptScreenCompleted = booleanPreferencesKey("TECHNICAL_OPT_SCREEN_COMPLETE")
 	private val quicEnabled = booleanPreferencesKey("QUIC_ENABLED")
 	private val isStreamingServerBannerDisplayed = booleanPreferencesKey("STREAMING_SERVER_DISPLAYED")
 	private val isPerAppSecurityBannerDisplayed = booleanPreferencesKey("DEFAULT_PER_APP_SECURITY_BANNER_DISPLAYED")
@@ -171,16 +172,24 @@ class DataStoreSettingsRepository(private val dataStoreManager: DataStoreManager
 		return dataStoreManager.getFromStore(statsDialogSkip) ?: Settings.FLAG_STATS_DIALOG_SKIP
 	}
 
+	override suspend fun setOnboardingCompleted() {
+		dataStoreManager.saveToDataStore(this.onboardingCompleted, true)
+	}
+
+	override suspend fun setTechnicalOptScreenCompleted() {
+		dataStoreManager.saveToDataStore(this.technicalOptScreenCompleted, true)
+	}
+
+	override suspend fun isOnboardingCompleted(): Boolean {
+		return dataStoreManager.getFromStore(onboardingCompleted) ?: Settings.FLAG_ONBOARDING_COMPLETED
+	}
+
+	override suspend fun isTechnicalOptScreenCompleted(): Boolean {
+		return dataStoreManager.getFromStore(technicalOptScreenCompleted) ?: Settings.FLAG_TECHNICAL_OPT_COMPLETED
+	}
+
 	override suspend fun setStatsDialogSkipped(skip: Boolean) {
 		dataStoreManager.saveToDataStore(this.statsDialogSkip, skip)
-	}
-
-	override suspend fun setWelcomeScreenCompleted() {
-		dataStoreManager.saveToDataStore(this.welcomeScreenCompleted, true)
-	}
-
-	override suspend fun isWelcomeScreenCompleted(): Boolean {
-		return dataStoreManager.getFromStore(welcomeScreenCompleted) ?: Settings.FLAG_WELCOME_SCREEN_COMPLETED
 	}
 
 	override suspend fun getQUICEnabled(): Boolean {
@@ -261,7 +270,8 @@ class DataStoreSettingsRepository(private val dataStoreManager: DataStoreManager
 						sentryEnabled = pref[sentryEnabled] ?: Settings.DEFAULT_SENTRY_ENABLED,
 						statsEnabled = pref[statsEnabled] ?: Settings.DEFAULT_STATS_ENABLED,
 						statsDialogSkip = pref[statsDialogSkip] ?: Settings.FLAG_STATS_DIALOG_SKIP,
-						welcomeScreenCompleted = pref[welcomeScreenCompleted] ?: Settings.FLAG_WELCOME_SCREEN_COMPLETED,
+						onboardingCompleted = pref[onboardingCompleted] ?: Settings.FLAG_ONBOARDING_COMPLETED,
+						technicalOptCompleted = pref[technicalOptScreenCompleted] ?: Settings.FLAG_TECHNICAL_OPT_COMPLETED,
 						quicEnabled = pref[quicEnabled] ?: Settings.DEFAULT_QUIC_ENABLED,
 						isStreamingServerBannerDisplayed = pref[isStreamingServerBannerDisplayed] ?: Settings.DEFAULT_STREAMING_SERVER_BANNER_DISPLAYED,
 						isPerAppSecurityBannerDisplayed = pref[isPerAppSecurityBannerDisplayed] ?: Settings.DEFAULT_PER_APP_SECURITY_BANNER_DISPLAYED,