+++ nym-vpn-android/app/src/main/java/net/nymtech/nymvpn/ui/common/buttons/surface/SurfaceSelectionGroupButton.kt
@@ -16,6 +16,7 @@ import androidx.compose.foundation.layout.offset
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.width
+import androidx.compose.foundation.layout.wrapContentWidth
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material3.Card
 import androidx.compose.material3.CardDefaults
@@ -81,7 +82,6 @@ fun SurfaceSelectionGroupButton(
 						modifier = Modifier
 							.weight(1f, false)
 							.padding(vertical = 8.dp.scaledHeight())
-							.fillMaxSize()
 							.padding(end = 4.dp.scaledWidth()),
 					) {
 						Box(modifier = Modifier.padding(start = anchorsPadding.scaledWidth()))
@@ -107,8 +107,8 @@ fun SurfaceSelectionGroupButton(
 						Box(
 							contentAlignment = Alignment.CenterEnd,
 							modifier = Modifier
-								.weight(0.35f)
-								.padding(horizontal = anchorsPadding.scaledWidth()),
+								.wrapContentWidth(Alignment.End)
+								.padding(start = 8.dp, end = anchorsPadding.scaledWidth()),
 						) {
 							trailing()
 						}