+++ nym-vpn-android/app/src/main/java/net/nymtech/nymvpn/ui/screens/settings/login/LoginUiState.kt
@@ -1,0 +1,7 @@
+package net.nymtech.nymvpn.ui.screens.settings.login
+
+data class LoginUiState(
+	val success: Boolean? = null,
+	val showMaxDevicesModal: Boolean = false,
+	val showTechnicalOptScreen: Boolean = false,
+)