+++ nym-vpn-android/CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Locale for high/low copy on Details screen (https://github.com/nymtech/nym-vpn-client/pull/4274)
 - Add reconnect on LAN bypass (https://github.com/nymtech/nym-vpn-client/pull/4282)
+- Add Onboarding flow (https://github.com/nymtech/nym-vpn-client/pull/4322)
+- Add technical opt screen after login/purchase (https://github.com/nymtech/nym-vpn-client/pull/4322)
 
 ### Changed
 - Replace mnemonic and access code with passphrase (https://github.com/nymtech/nym-vpn-client/pull/4274)
@@ -21,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Changes for Custom DNS screen reconnecting logic (https://github.com/nymtech/nym-vpn-client/pull/4282)
 - Remove reconnect modals for QUIC on Censorship screen, add snack bar and debounce (https://github.com/nymtech/nym-vpn-client/pull/4282)
 - Promote VPN to foreground immediately at service entry (https://github.com/nymtech/nym-vpn-client/pull/4291)
+- UI changes for User welcome screen (https://github.com/nymtech/nym-vpn-client/pull/4322)
 
 ### Fixed
 - "Add" button on DNS screen now expand to the left for verbose languages (https://github.com/nymtech/nym-vpn-client/pull/4277)